split off modules that are specifically to do with the secondary window into their own folder
