+++ src/secondary-window/shims/block-unnecessary-notifications.ts
