+++ src/secondary-window/shims/block-unnecessary-ui.ts
