+++ src/secondary-window/shims/disable-canvas.ts
