+++ src/utils/foundry.ts
@@ -8,6 +8,7 @@ export function getGame() {
   return game as Game;
 }
 
+// TODO fix this to be a documentsheet?
 export function getEntitySheet(
   id: string,
   type: EntityType
